+++ Examples/iOSExample/ViewController.swift
@@ -26,9 +26,9 @@ class ViewController: UIViewController, UICollectionViewDataSource {
       ("Adjusted", mainColor.adjustedHue(amount: 45)),
       ("Complemented", mainColor.complemented()),
       ("Inverted", mainColor.inverted()),
-      ("Mix Blue", mainColor.mixed(color: .blue)),
-      ("Mix Green", mainColor.mixed(color: .green)),
-      ("Mix Yellow", mainColor.mixed(color: .yellow)),
+      ("Mix Blue", mainColor.mixed(withColor: .blue)),
+      ("Mix Green", mainColor.mixed(withColor: .green)),
+      ("Mix Yellow", mainColor.mixed(withColor: .yellow)),
       ("Tinted", mainColor.tinted()),
       ("Shaded", mainColor.shaded())
     ]