+++ Sources/DynamicColor.swift
@@ -126,7 +126,7 @@ public extension DynamicColor {
    - parameter hexString: A hexa-decimal color number representation to be compared to the receiver.
    - returns: true if the receiver and the string are equals, otherwise false.
    */
-  public func isEqualToHexString(_ hexString: String) -> Bool {
+  public func isEqual(toHexString hexString: String) -> Bool {
     return self.toHexString() == hexString
   }
 
@@ -136,7 +136,7 @@ public extension DynamicColor {
    - parameter hex: A UInt32 that represents the hexa-decimal color.
    - returns: true if the receiver and the integer are equals, otherwise false.
    */
-  public func isEqualToHex(_ hex: UInt32) -> Bool {
+  public func isEqual(toHex hex: UInt32) -> Bool {
     return self.toHex() == hex
   }
 
@@ -167,17 +167,16 @@ public extension DynamicColor {
    - parameter weight: The weight specifies the amount of the given color object (between 0 and 1). The default value is 0.5, which means that half the given color and half the receiver color object should be used. 0.25 means that a quarter of the given color object and three quarters of the receiver color object should be used.
    - returns: A color object corresponding to the two colors object mixed together.
    */
-  public final func mixed(color: DynamicColor, weight: CGFloat = 0.5) -> DynamicColor {
+  public final func mixed(withColor color: DynamicColor, weight: CGFloat = 0.5) -> DynamicColor {
     let normalizedWeight = clip(weight, 0, 1)
-    let remains          = 1 - normalizedWeight
 
     let c1 = toRGBAComponents()
     let c2 = color.toRGBAComponents()
 
-    let red   = remains * c1.r + normalizedWeight * c2.r
-    let green = remains * c1.g + normalizedWeight * c2.g
-    let blue  = remains * c1.b + normalizedWeight * c2.b
-    let alpha = remains * c1.a + normalizedWeight * c2.a
+    let red   = c1.r + normalizedWeight * (c2.r - c1.r)
+    let green = c1.g + normalizedWeight * (c2.g - c1.g)
+    let blue  = c1.b + normalizedWeight * (c2.b - c1.b)
+    let alpha = c1.a + normalizedWeight * (c2.a - c1.a)
 
     return DynamicColor(red: red, green: green, blue: blue, alpha: alpha)
   }
@@ -189,7 +188,7 @@ public extension DynamicColor {
    - returns: A lighter DynamicColor.
    */
   public final func tinted(amount: CGFloat = 0.2) -> DynamicColor {
-    return mixed(color: DynamicColor.white, weight: amount)
+    return mixed(withColor: .white, weight: amount)
   }
 
   /**
@@ -199,6 +198,6 @@ public extension DynamicColor {
    - returns: A darker DynamicColor.
    */
   public final func shaded(amount: CGFloat = 0.2) -> DynamicColor {
-    return mixed(color: DynamicColor(red:0, green:0, blue: 0, alpha:1), weight: amount)
+    return mixed(withColor: DynamicColor(red:0, green:0, blue: 0, alpha:1), weight: amount)
   }
 }